backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic-settings==2.9.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# AI & ML Integrations
# Use OpenAI SDK for Friendli Serverless compatibility
openai==1.40.2
langchain==0.0.350
langchain-community==0.0.3

# AWS SDK and Strands Agent SDK
boto3==1.34.82
botocore==1.34.82

strands-agents==1.0.0

# AWS Observability and Security
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation==0.45b0

# Vector Database
weaviate-client==3.25.3

# Graph Database
neo4j==5.14.1

# Redis
redis==5.0.1
celery==5.3.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==1.1.0

# HTTP Clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0

# File Storage
boto3==1.34.0
minio==7.2.0

# Email
fastapi-mail==1.4.1

# Monitoring & Logging
structlog==23.2.0
loguru==0.7.2
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment Management
python-dotenv==1.0.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Validation & Serialization
marshmallow==3.20.1
pydantic[email]==2.5.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Audio Processing (for VAPI integration)
pydub==0.25.1
speech-recognition==3.10.0

# Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Configuration
dynaconf==3.2.4

# Health Checks
healthcheck==1.3.3

# API Documentation
fastapi-limiter==0.1.6
slowapi==0.1.9
